Turn the maximum length ifs into asserts
I think the ifs were a bit misleading in that they suggested that counter
values this big were a normal part of operation. However, the main reason
for these checks is to ensure a theoretical bound on the construction
which no one should ever run into. The secondary reason for why they are
misleading is that they suggest that encrypting this much data under a
single key is cryptographically safe, which I haven't actually checked.
(I think XSalsa20 is relatively uncomplicated in this regard, but again,
I don't know for sure. If we end up switching to AES-SIV, though, this
kind of thing definitely becomes a concern.)